requirements.txt
amqp==5.0.6
asgiref==3.4.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.5.30
charset-normalizer==2.0.6
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.7
django-filter==21.1
djangorestframework==3.12.4
docopt==0.6.2
idna==3.2
kombu==5.1.0
Markdown==3.3.4
pipreqs==0.4.10
prompt-toolkit==3.0.20
Pygments==2.10.0
pytz==2021.1
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
yarg==0.1.9
